// admin/src/api/generated/schemas/adminUpdateSwapConfigDetailBody.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Internal Wallet Admin API
 * Admin API (served via api-gateway)
 * OpenAPI spec version: v1
 */

export type AdminUpdateSwapConfigDetailBody = {
  provider_name?: string;
  chain_name?: string;
  chain_symbol?: string;
  token_name?: string;
  icon_url?: string;
  is_enabled?: boolean;
  priority?: number;
  router_address?: string;
  min_swap_amount_usd?: string;
  max_swap_amount_usd?: string;
  default_slippage?: string;
  max_slippage?: string;
  fee_rate?: string;
  config_json?: string;
};